o.c.diag.pvfields: PVField can update

+++ applications/plugins/org.csstudio.diag.pvfields/src/org/csstudio/diag/pvfields/Preferences.java
@@ -5,8 +5,8 @@ public class Preferences
 	/** Timeout for querying {@link DataProvider}s
 	 *  @return Milliseconds
 	 */
-	public static long getTimeout()
+	public static int getTimeout()
 	{
-		return 60*1000L;
+		return 60*1000;
 	}
 }
